Test with Ruby 2.4, fix warnings
