+++ .travis.yml
@@ -6,6 +6,7 @@ rvm:
   - 2.1.8
   - 2.2.3
   - 2.3.0
+  - 2.4.1
 gemfile:
   - gemfiles/rails3.2.gemfile
   - gemfiles/rails4.1.gemfile
