+++ lib/stronger_parameters/constraints/boolean_constraint.rb
@@ -1,4 +1,4 @@
-require 'stronger_parameters/constraints'
+require 'stronger_parameters/constraint'
 
 module StrongerParameters
   class BooleanConstraint < Constraint
