+++ lib/stronger_parameters/constraints/comparison_constraints.rb
@@ -1,4 +1,4 @@
-require 'stronger_parameters/constraints'
+require 'stronger_parameters/constraint'
 
 module StrongerParameters
   class ComparisonConstraints < Constraint
