+++ lib/stronger_parameters/constraints/date_time_constraint.rb
@@ -1,4 +1,4 @@
-require 'stronger_parameters/constraints'
+require 'stronger_parameters/constraint'
 
 module StrongerParameters
   class DateTimeConstraint < Constraint
