+++ lib/stronger_parameters/constraints/nil_string_constraint.rb
@@ -1,4 +1,4 @@
-require 'stronger_parameters/constraints'
+require 'stronger_parameters/constraint'
 
 module StrongerParameters
   class NilStringConstraint < Constraint
