+++ test/decimal_constraint_test.rb
@@ -7,27 +7,27 @@ describe 'decimal parameter constraints' do
     permits '1.23', as: 1.23
     permits '-1.23', as: -1.23
     permits 1.23
-    permits -1.23
+    permits(-1.23)
 
     permits '12.34', as: 12.34
     permits '-12.34', as: -12.34
     permits 12.34
-    permits -12.34
+    permits(-12.34)
 
     permits 1
-    permits -1
+    permits(-1)
     permits '1', as: 1
     permits '-1', as: -1
 
     permits '1.2', as: 1.2
     permits '-1.2', as: -1.2
     permits 1.2
-    permits -1.2
+    permits(-1.2)
 
     rejects '123', as: 123
     rejects '-123', as: -123
     rejects 123
-    rejects -123
+    rejects(-123)
 
     rejects 'abc'
     rejects '1.2.3'
@@ -38,17 +38,17 @@ describe 'decimal parameter constraints' do
     rejects '1234', as: 1234
     rejects '-1234', as: -1234
     rejects 1234
-    rejects -1234
+    rejects(-1234)
 
     rejects '1.234', as: 1.234
     rejects '-1.234', as: -1.234
     rejects 1.234
-    rejects -1.234
+    rejects(-1.234)
 
     rejects '123.45', as: 123.45
     rejects '-123.45', as: 123.45
     rejects 123.45
-    rejects -123.45
+    rejects(-123.45)
   end
 
   describe 'with 0 scale and 5 precision' do
@@ -56,12 +56,12 @@ describe 'decimal parameter constraints' do
 
     permits 1
     permits '1', as: 1
-    permits -1
+    permits(-1)
     permits '-1', as: -1
 
     permits 12345
     permits '12345', as: 12345
-    permits -12345
+    permits(-12345)
     permits '-12345', as: -12345
 
     rejects 'abc'
@@ -73,11 +73,11 @@ describe 'decimal parameter constraints' do
     rejects '1.23', as: 1.23
     rejects '-1.23', as: -1.23
     rejects 1.23
-    rejects -1.23
+    rejects(-1.23)
 
     rejects 123456
     rejects '123456', as: 123456
-    rejects -123456
+    rejects(-123456)
     rejects '-123456', as: -123456
   end
 end
